requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
httpx==0.24.1
python-jose==3.3.0
python-multipart==0.0.6
pycryptodome==3.19.0
python-dotenv==1.0.0
pytest==7.4.2 